+++ actiontext/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Rename `rich_text_area` methods into `rich_textarea`
+
+    Old names are still available as aliases.
+
+    *Sean Doyle*
+
+
 *   Only sanitize `content` attribute when present in attachments.
 
     *Petrik de Heus*